Create ChatList component
